requirements.txt
matplotlib==2.1.1
numpy==1.13.3
Pillow==4.3.0

opencv-python==3.4.0.12
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
olefile==0.44
backports.functools-lru-cache==1.4
cycler==0.10.0
six==1.11.0